+++ pandamesh/triangle_mesher.py
@@ -10,13 +10,14 @@ from pandamesh.common import (
     check_geodataframe,
     repr,
     separate_geodataframe,
+    to_geodataframe,
+    to_ugrid,
 )
-from pandamesh.mesher_base import MesherBase
 from pandamesh.triangle_enums import DelaunayAlgorithm
 from pandamesh.triangle_geometry import collect_geometry, polygon_holes
 
 
-class TriangleMesher(MesherBase):
+class TriangleMesher:
     """
     Wrapper for the python bindings to Triangle. This class must be initialized
     with a geopandas GeoDataFrame containing at least one polygon, and a column
@@ -213,3 +214,23 @@ class TriangleMesher(MesherBase):
         vertices[:, 0] += self._xoff
         vertices[:, 1] += self._yoff
         return vertices, result["triangles"]
+
+    def generate_geodataframe(self) -> gpd.GeoDataFrame:
+        """
+        Generate a mesh and return it as a geopandas GeoDataFrame.
+
+        Returns
+        -------
+        mesh: geopandas.GeoDataFrame
+        """
+        return to_geodataframe(*self.generate())
+
+    def generate_ugrid(self) -> "xugrid.Ugrid2d":  # type: ignore # noqa  pragma: no cover
+        """
+        Generate a mesh and return it as an xugrid Ugrid2d.
+
+        Returns
+        -------
+        mesh: xugrid.Ugrid2d
+        """
+        return to_ugrid(*self.generate())